pyproject.toml
[tool.poetry]
name = "nonebot-plugin-dice-narrator"
version = "0.1.0"
description = "一只可爱的AI掷骰姬"
authors = ["KroMiose <li_xiangff@163.com>"]
readme = "README.md"
packages = [{ include = "nonebot_plugin_dice_narrator" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
openai = "^1.13.3"
nonebot-adapter-onebot = "^2.4.2"
miose-toolkit-common = "^0.1.2"
nonebot2 = { extras = ["fastapi"], version = "^2.2.1" }
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
nb-cli = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "nonebot-plugin-dice-narrator"
version = "0.1.0"
description = "一只可爱的AI掷骰姬"
readme = "README.md"
requires-python = ">=3.9,<3.12"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
plugins = ["nonebot_plugin_dice_narrator"]
plugin_dirs = []
builtin_plugins = []

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
priority = "primary"

[tool.ruff]
extend-exclude = []
ignore = [
    "E402",
    "E501",
    "E722",
    "B008",
    "B905",
    "F401",
    "TID252",
    "TRY002",
    "TRY003",
    "RUF001",
    "RUF002",
    "RUF003",
    "RUF006",
    "RUF012",
    "RUF100",
    "PGH003",
    "N801",
    "A003",
]
select = [
    "F",
    "E",
    "I",
    "B",
    "A",
    "COM",
    "C4",
    "ISC",
    "PIE",
    "PYI",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SIM",
    "TID",
    "ARG",
    "PTH",
    "PGH",
    "TRY",
    "RUF",
]
